Add `SpriteAnimationTicker` - Move ticking logic out of `SpriteAnimation`
